Fixing ShowMessage type serialization

+++ org.jboss.tools.language-server.messages/src/main/java/org/jboss/tools/lsp/messages/ShowMessageParams.java
@@ -14,7 +14,7 @@ public class ShowMessageParams {
      */
     @SerializedName("type")
     @Expose
-    private Double type;
+    private Integer type;
     /**
      * The actual message
      * 
@@ -29,7 +29,7 @@ public class ShowMessageParams {
      * @return
      *     The type
      */
-    public Double getType() {
+    public Integer getType() {
         return type;
     }
 
@@ -39,11 +39,11 @@ public class ShowMessageParams {
      * @param type
      *     The type
      */
-    public void setType(Double type) {
+    public void setType(Integer type) {
         this.type = type;
     }
 
-    public ShowMessageParams withType(Double type) {
+    public ShowMessageParams withType(int type) {
         this.type = type;
         return this;
     }
